Use a direct address for P4 rather than a pointer

+++ src/kernel/memory/pmm.c
@@ -1,25 +1,25 @@
 #include <memory.h>
 
-uintptr_t first = 0;
+uint64_t first = 0;
 
-void pmm_free(uintptr_t page)
+void pmm_free(uint64_t page)
 {
-  page = (uintptr_t)P2V(page);
-  *(uintptr_t *)page = first;
+  page = (uint64_t)P2V(page);
+  *(uint64_t *)page = first;
   first = page;
 }
 
-uintptr_t pmm_alloc()
+uint64_t pmm_alloc()
 {
-  uintptr_t page = first;
-  first = page?*(uintptr_t *)page:0;
-  page = (uintptr_t)(page?V2P(page):0);
+  uint64_t page = first;
+  first = page?*(uint64_t *)page:0;
+  page = (uint64_t)(page?V2P(page):0);
   return page;
 }
 
-uintptr_t pmm_calloc()
+uint64_t pmm_calloc()
 {
-  uintptr_t page = pmm_alloc();
+  uint64_t page = pmm_alloc();
   memset(P2V(page), 0, 0x1000);
   return page;
 }
